requirements.txt
openai== 1.37.1
azure.cognitiveservices.speech==1.38.0
python-dotenv== 1.0.1
semantic-kernel==1.0.5
asyncio==3.4.3
pyodbc==5.1.0
faker==26.0.0
